Extract Routing building logic into a separate class
